перенес тест в компонент, удалил лишние сторисы
