+++ src/gameOfLife/components/CellComponent.test.tsx
@@ -47,21 +47,3 @@ describe('CellComponent tests', () => {
 
 });
 
-describe('BoardComponent tests', () => {
-    it('show cell component with id', () => {
-        
-        render(<BoardComponent/>);
-        const allElements = screen.getAllByText('2399');
-
-        expect(allElements).toHaveLength(1);
-    });
-
-    it('count cell element equal 2500', () => {
-        const {container} = render(<BoardComponent/>);
-        const allElements = container.getElementsByClassName("cell")
-
-        expect(allElements).toHaveLength(2500);
-    });
-
-    
-});
